backend/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0.
anyio==4.8.0
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Django==5.1.7
django-environ==0.12.0
django-recaptcha==4.0.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.64.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
psycopg2-binary==2.9.5
whitenoise==6.9.0
django-cors-headers==4.3.1
cryptography>=36.0.0
pycryptodome>=3.22.0